docker/substack-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-substack==0.1.17
pydantic==2.9.2
